Implemented and Tested User and Authenticator Model Classes
Implemented and Tested two new model classes to provide business logic for logging into the system. User class models a user, with a name, password, and an arraylist of roles, from a newly created Role enum. The authenticator class stores the list of users in the system, with two example users being initialized in the constructor. Authenticator has two methods, one for authenticating a user's provided credentials and another to be used for getting the roles from a user stored in the list.