Implemented and Tested Compartment and Inventory classes
Implemented and Tested two new classes that define the storage compartments for different types of special quality stock items. The compartment itself holds a specific capacity of items of a specific type or combination of types. The inventory class holds the list of these compartments and manages all of them. Also refactored the name of an unrelated test case that I mispelt earlier.