Updated View, Code Behind, and FXML file to support the login page functionality
Updated FXML file, view model, and view to handle the responsibilities necessary for submitting user credentials to log into the system. Deleted the old example files from the starter project as well. On the login page, the submit button is disabled unless a valid, non-blank, username and password combination are entered. Currently, the system pops up an alert if the login was unsuccessfully due to the credentials not being present in the system.