Implemented new FXML, ViewModel, and View for the landing page
Updated LoginCodeBehind to pop up the new landing page window once a user makes a successful login attempt. The landing page says hello to the user that logged in and shows them their roles. The add stock and view changes buttons are enabled and disabled automatically depending on the roles of the user that logged into the system.