Implemented view and Model for adding stock page and for the page to list the compartments available to add stock to
Now, when the add stock button on the landing page is selected, the user will be prompted to enter details for adding a new stock. The user will not be able to add the stock unless the details are filled out correctly. If there is available space in a compartment for the stock the user wants to add, a new window will pop up with a list of compartments that they may add the stock to. A few model classes were altered as necessary to accomplish this task, with new test cases being written accordingly.