Feature-UpdateAddStockPageToIncludeSelectionOfStockCategory
Updated fxml, view, and view model classes that relate to the process of adding a new stock to the system to support functionality for allowing crewmates to select the category (other, food, or munitions) for the stock that they are adding. Also ensured that the updated classes maintain test coverage.